requirements.txt
openai>=1.54.4
python-telegram-bot>=21.7
fastapi>=0.115.2
uvicorn>=0.32.0
